+++ packages/office-ui-fabric-react/src/components/Checkbox/examples/Checkbox.ImplementationExamples.tsx
@@ -28,12 +28,6 @@ export class CheckboxImplementationExamples extends React.Component<{}, ICheckbo
   public render() {
     let { isChecked } = this.state;
 
-    let styles: ICheckboxStyles = {
-      root: {
-        marginTop: '10px'
-      }
-    };
-
     return (
       <div>
         <Checkbox
@@ -43,7 +37,6 @@ export class CheckboxImplementationExamples extends React.Component<{}, ICheckbo
             onFocus: () => { console.log('Uncontrolled checkbox is focused'); },
             onBlur: () => { console.log('Uncontrolled checkbox is blured'); }
           } }
-          styles={ styles }
           ariaDescribedBy={ 'descriptionID' }
         />
         <label id='descriptionID' className='screenReaderOnly'>Uncontroller checkbox description</label>
@@ -52,13 +45,11 @@ export class CheckboxImplementationExamples extends React.Component<{}, ICheckbo
           label='Uncontrolled checkbox with defaultChecked true'
           defaultChecked={ true }
           onChange={ this._onCheckboxChange }
-          styles={ styles }
         />
 
         <Checkbox
           label='Disabled uncontrolled checkbox'
           disabled={ true }
-          styles={ styles }
         />
 
         <Checkbox
@@ -66,25 +57,21 @@ export class CheckboxImplementationExamples extends React.Component<{}, ICheckbo
           disabled={ true }
           defaultChecked={ true }
           onChange={ this._onCheckboxChange }
-          styles={ styles }
         />
 
         <Checkbox
           label='Controlled checkbox'
           checked={ isChecked }
           onChange={ this._onControlledCheckboxChange }
-          styles={ styles }
         />
 
         <Checkbox
           label='Checkbox rendered with boxSide "end"'
           boxSide='end'
-          styles={ styles }
         />
 
         <Checkbox
           label='Persona Checkbox'
-          styles={ styles }
           onRenderLabel={ this._renderPersonaLabel }
         />
       </div>
