+++ packages/office-ui-fabric-react/src/components/Dropdown/Dropdown.tsx
@@ -464,7 +464,7 @@ export class Dropdown extends BaseComponent<IDropdownInternalProps, IDropdownSta
     let { selectedIndices = [] } = this.state;
     let id = this._id;
     let isItemSelected = item.index !== undefined && selectedIndices ? selectedIndices.indexOf(item.index) > -1 : false;
-    let checkboxStyles = getCheckboxStyles(getTheme());
+    // let checkboxStyles = getCheckboxStyles(getTheme());
 
     return (
       !this.props.multiSelect ?
@@ -495,7 +495,7 @@ export class Dropdown extends BaseComponent<IDropdownInternalProps, IDropdownSta
         ) : (
           <Checkbox
             id={ id + '-list' + item.index }
-            ref={ Dropdown.Option + item.index }
+            // ref={ Dropdown.Option + item.index }
             key={ item.key }
             data-index={ item.index }
             data-is-focusable={ true }
@@ -516,13 +516,13 @@ export class Dropdown extends BaseComponent<IDropdownInternalProps, IDropdownSta
             role='option'
             aria-selected={ isItemSelected ? 'true' : 'false' }
             checked={ isItemSelected }
-            // Hover is being handled by focus styles
-            // so clear out the explicit hover styles
-            styles={ {
-              checkboxHovered: checkboxStyles.checkbox,
-              checkboxCheckedHovered: checkboxStyles.checkboxChecked,
-              textHovered: checkboxStyles.text
-            } }
+          // Hover is being handled by focus styles
+          // so clear out the explicit hover styles
+          // styles={ {
+          //   checkboxHovered: checkboxStyles.checkbox,
+          //   checkboxCheckedHovered: checkboxStyles.checkboxChecked,
+          //   textHovered: checkboxStyles.text
+          // } }
           >{ onRenderOption(item, this._onRenderOption) }
           </Checkbox>
         )
